Rename template package and app id to demo.users
