+++ app/src/androidTest/java/dev/burnoo/demo/listapp/ComposeTest.kt
@@ -1,9 +1,9 @@
-package dev.burnoo.template
+package dev.burnoo.demo.listapp
 
 import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
-import dev.burnoo.template.composable.AppRouter
-import dev.burnoo.template.core.designsystem.AppTheme
+import dev.burnoo.demo.listapp.composable.AppRouter
+import dev.burnoo.demo.listapp.core.designsystem.AppTheme
 import org.junit.Rule
 import org.junit.Test
