+++ core/designsystem/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="dev.burnoo.template.ui.theme">
+    package="dev.burnoo.demo.listapp.ui.theme">
 
     <application android:theme="@style/Theme.Material3.Light.NoActionBar" />
 </manifest>
