+++ feature/home/src/main/java/dev/burnoo/demo/listapp/home/HomeRouter.kt
@@ -1,4 +1,4 @@
-package dev.burnoo.template.feature.home
+package dev.burnoo.demo.listapp.home
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
